// frontend/src/types/backend/clinic-settings-dto.ts
/**
 * This is a TypeGen auto-generated file.
 * Any changes made to this file can be lost when this file is regenerated.
 */

export class ClinicSettingsDto {
  timeZone: string = "UTC";
  currency: string = "USD";
  dateFormat: string = "dd/MM/yyyy";
  locale: string = "en-US";
  enablePatientPortal: boolean;
}